Activation Functions
I did the activation functions here, and going to get to the backward passes once I've finished all the forwards for everything I need like loss functions and optimizer and stuff. I'm currently making the softmax activation, but wanted to mess around with something, and if it doesn't work I want a backup to go back to so this is that.