Update MNIST_Fashion.py
Did some optimizations to the loss class, added a general loss class just in case I want to try a different loss type later (chasing that oh so delicious scalability lol) and added average loss, fixed loss so it works with batches, and added a choice frequency dictionary and a pecentage times that a specific item was chosen, just to see how well the neural net is doing and what its choosing. (I think somehting might be wrong, it guesses like one or two things most of the time, i think it is just the nature of the random initalization of my weights matrix ¯\_(ツ)_/¯. This neural network is getting pretty big and complex and I'm rather proud of it's progress so far! :)