Most recent code
Still not quite able to figure out why it is predicting the same thing consistently, I am considering abandoning this one and working on some more impactful projects. I've asked several people in the ML space and they don't know either (and thats not even includinng 50ish people on stack overflow). IDK honestly but I have some more important and impactful project ideas on a nice list that I'd rather make, and this project (along with the NNFS book ofc) served its purpose and I understand Neural Nets to some degree of competency where I feel like I could effectivly leverage pytorch or tf and plt to make some important, impactful, and useful things. If I can't figure it out by the end of the month I'm scrapping it. I have a few things I want to try like leaky ReLu or maybe a different hidden layer activation function all together, also maybe trying random data to see if the problem is my neural net or the data (probably not the data tho its MNIST lol), and maybe increasing my floating point presicion b/c I feel like that might be why the argmax of the final softmax calc is mostly the same for one batch. I though dropout would help bc I felt like it was happening becuase of a few neurons doing a bad and occasionally they would be dropped out with dropout, but that wasnt the case. I'm of course new to the field so I imagine my lack of experince is a factor in my inablitiy to solve this one alone, i dont know. I hate to scrap projects but this one is consuming more time than I would have liked. Tl;DR: I might scrap this one. Wow this was wayyyyy longer than I meant it to be but its been a while since I've updated so I thought a longer-form note was in order. Anyway this may be one of the last posts to this project so thanks for checking it out!